+++ frontend/src/lib/i18n.js
@@ -1,56 +1,45 @@
-// Tiny dependency-free i18n. English source strings are the keys; locale files in
-// src/locales/ map them to translations and are lazy-loaded (Vite code-splits each
-// import.meta.glob entry), so the initial bundle stays English-only.
-// Exercise instructions come from separately generated packs in src/instr/ (one per
-// language, from the upstream dataset) — also lazy-loaded on language switch.
+// Browser-only shell of the i18n module.
+//
+// Pure state, constants and readers (t, dateLocale, instrFor, getLang, LANGS, INSTR_LANGS)
+// live in i18n-core.js so plain Node can reuse the training logic without a Vite environment.
+// This file adds the two pieces that genuinely need the browser: the async `setLang` (which
+// lazy-loads locale packs via import.meta.glob) and the React subscription hook `useLang`
+// (which re-renders components on language switch).
+
 import { useSyncExternalStore } from 'react'
+import {
+  LANGS, INSTR_LANGS, DATE_LOCALES,
+  getLang, dateLocale, t, instrFor, getVersion, _setLangState
+} from './i18n-core.js'
 
-// UI languages. de/pt have no instruction pack upstream — instructions fall back to English.
-export const LANGS = {
-  en: 'English', de: 'Deutsch', es: 'Español', fr: 'Français', it: 'Italiano',
-  pt: 'Português', pl: 'Polski', tr: 'Türkçe', ru: 'Русский', zh: '中文',
-  ko: '한국어', hi: 'हिन्दी'
-}
-export const INSTR_LANGS = ['en', 'es', 'fr', 'it', 'tr', 'ru', 'zh', 'hi', 'pl', 'ko']
-const DATE_LOCALES = {
-  en: 'en-GB', de: 'de-DE', es: 'es-ES', fr: 'fr-FR', it: 'it-IT', pt: 'pt-PT',
-  pl: 'pl-PL', tr: 'tr-TR', ru: 'ru-RU', zh: 'zh-CN', ko: 'ko-KR', hi: 'hi-IN'
-}
+// Re-export everything verbatim so existing callers (`import { t } from '../lib/i18n.js'`)
+// keep working unchanged — the split is invisible at the import site.
+export { LANGS, INSTR_LANGS, DATE_LOCALES, getLang, dateLocale, t, instrFor }
 
+// Vite code-splits each locale pack into its own chunk via import.meta.glob; instructions use
+// the same mechanism in src/instr/. Both are lazy, so the production bundle ships English only.
 const localePacks = import.meta.glob('../locales/*.js')
 const instrPacks = import.meta.glob('../instr/*.js')
 
-let lang = 'en'
-let dict = {}
-let instr = null            // { exId: [steps] } for the current language, null = English
-let version = 0
+// React subscription bookkeeping — kept here, not in core, so core has zero React coupling.
 const subs = new Set()
-const notify = () => { version++; subs.forEach(f => f()) }
-
-export const getLang = () => lang
-export const dateLocale = () => DATE_LOCALES[lang] || 'en-GB'
-
-// Translate a source string; {0},{1}… are replaced with args (also on the English fallback).
-export function t(s, ...args) {
-  let v = dict[s] || s
-  for (let i = 0; i < args.length; i++) v = v.replaceAll('{' + i + '}', args[i])
-  return v
-}
-// Instructions for an exercise in the current language (English steps as fallback).
-export const instrFor = ex => (instr && instr[ex.id]) || ex.st || []
+const notify = () => { subs.forEach(f => f()) }
 
 export async function setLang(l) {
   if (!LANGS[l]) l = 'en'
-  if (l === lang && version > 0) return
-  lang = l
+  // _setLangState mutates core state and bumps the version; we read it back for the
+  // subscription selector below.
+  if (l === getLang() && getVersion() > 0) return
+  let dict = {}, instr = null
   try {
     dict = l === 'en' ? {} : (await localePacks['../locales/' + l + '.js']()).default
     instr = l === 'en' || !INSTR_LANGS.includes(l) ? null : (await instrPacks['../instr/' + l + '.js']()).default
   } catch (e) { dict = {}; instr = null }
+  _setLangState(l, dict, instr)
   notify()
 }
 
 // Re-renders the subscribing component (and its children) whenever the language changes.
 export function useLang() {
-  return useSyncExternalStore(fn => { subs.add(fn); return () => subs.delete(fn) }, () => version)
+  return useSyncExternalStore(fn => { subs.add(fn); return () => subs.delete(fn) }, getVersion)
 }
